PokemonAPI/pyproject.toml
[tool.poetry]
name = "pokemon api"
version = "0.0.1"
description = ""
authors = ["NEBIE Guillaume <nebie.guillaume.lale@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
packages = [{ include = "app" }]


[tool.poetry.scripts]
app = "app.src.__main__:main"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
httpx = "^0.27.0"
pydantic = "^2.7"
sqlmodel = "^0.0.16"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
black = "^24.4"
ruff = "^0.4"
bandit = "^1.7"
pytest-django = "^4.11.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
